subscriber: Move common test utility to fmt/mod.rs
The `MockWriter` and `MockMakeWriter` are used in both writer.rs and
format.rs tests. Move them to a common test module so that they
can be used anywhere else.

Fixes: #300
