+++ tracing-subscriber/src/fmt/format.rs
@@ -509,3 +509,88 @@ impl<'a> fmt::Display for FmtLevel<'a> {
         }
     }
 }
+
+#[cfg(test)]
+mod test {
+
+    use crate::fmt::test::MockWriter;
+    use crate::fmt::time::FormatTime;
+    use lazy_static::lazy_static;
+    use tracing::{self, subscriber::with_default};
+
+    use std::fmt;
+    use std::sync::Mutex;
+
+    struct MockTime;
+    impl FormatTime for MockTime {
+        fn format_time(&self, w: &mut dyn fmt::Write) -> fmt::Result {
+            write!(w, "fake time")
+        }
+    }
+
+    #[cfg(feature = "ansi")]
+    #[test]
+    fn with_ansi_true() {
+        lazy_static! {
+            static ref BUF: Mutex<Vec<u8>> = Mutex::new(vec![]);
+        }
+
+        let make_writer = || MockWriter::new(&BUF);
+        let expected = "\u{1b}[2mfake time\u{1b}[0m\u{1b}[32m INFO\u{1b}[0m tracing_subscriber::fmt::format::test: some ansi test\n";
+        test_ansi(make_writer, expected, true, &BUF);
+    }
+
+    #[cfg(feature = "ansi")]
+    #[test]
+    fn with_ansi_false() {
+        lazy_static! {
+            static ref BUF: Mutex<Vec<u8>> = Mutex::new(vec![]);
+        }
+
+        let make_writer = || MockWriter::new(&BUF);
+        let expected = "fake time INFO tracing_subscriber::fmt::format::test: some ansi test\n";
+
+        test_ansi(make_writer, expected, false, &BUF);
+    }
+
+    #[cfg(not(feature = "ansi"))]
+    #[test]
+    fn without_ansi() {
+        lazy_static! {
+            static ref BUF: Mutex<Vec<u8>> = Mutex::new(vec![]);
+        }
+
+        let make_writer = || MockWriter::new(&BUF);
+        let expected = "fake time INFO tracing_subscriber::fmt::format::test: some ansi test\n";
+        let subscriber = crate::fmt::Subscriber::builder()
+            .with_writer(make_writer)
+            .with_timer(MockTime)
+            .finish();
+
+        with_default(subscriber, || {
+            tracing::info!("some ansi test");
+        });
+
+        let actual = String::from_utf8(BUF.try_lock().unwrap().to_vec()).unwrap();
+        assert_eq!(expected, actual.as_str());
+    }
+
+    #[cfg(feature = "ansi")]
+    fn test_ansi<T>(make_writer: T, expected: &str, is_ansi: bool, buf: &Mutex<Vec<u8>>)
+    where
+        T: crate::fmt::MakeWriter + Send + Sync + 'static,
+    {
+        let subscriber = crate::fmt::Subscriber::builder()
+            .with_writer(make_writer)
+            .with_ansi(is_ansi)
+            .with_timer(MockTime)
+            .finish();
+
+        with_default(subscriber, || {
+            tracing::info!("some ansi test");
+        });
+
+        let actual = String::from_utf8(buf.try_lock().unwrap().to_vec()).unwrap();
+        assert_eq!(expected, actual.as_str());
+    }
+}
